+++ app/value_objects/workbaskets/create_measures/validation_helpers/candidates_validations_summarizer.rb
