+++ app/value_objects/workbaskets/create_measures/validation_helpers/excluded_geographical_areas.rb
