+++ app/value_objects/workbaskets/create_measures/validation_helpers/footnotes.rb
